Added verification access of order

+++ OrdersHandler/src/AppBundle/Controller/RepairOrderController.php
@@ -15,6 +15,15 @@ class RepairOrderController extends Controller
         $em = $this->getDoctrine()->getManager();
 
         $entities = $em->getRepository('AppBundle:RepairOrder')->findAll();
+        foreach ($entities as $entity) {
+            try {
+                $this->denyAccessUnlessGranted('view', $entity, 'Access denied!');
+            } catch (AccessDeniedException $ex) {
+                $key = array_search($entity, $entities);
+                unset($entities[$key]);
+                $entities = array_values($entities);
+            }
+        }
 
         return $this->render('AppBundle:RepairOrder:index.html.twig', array(
                 'entities' => $entities
@@ -27,14 +36,23 @@ class RepairOrderController extends Controller
         $user = $this->getUser();
 
         $entity = new RepairOrder();
-        $this->denyAccessUnlessGranted('view', $entity, 'Unauthorized access!');
+        try {
+            $this->denyAccessUnlessGranted('create', $entity, 'Access denied!');
+        } catch (AccessDeniedException $ex) {
+            return $this->render('default/index.html.twig', array(
+                    'errorMessages' => [
+                        $ex->getMessage()
+                    ]
+                )
+            );
+        }
         $entity->setStatus(RepairOrderType::STATUS_OPEN);
         $entity->setUser($user);
 
         $form = $this->createForm($this->get('form.order.type'), $entity, array(
             'action' => $this->generateUrl('repairorder_new'),
         ));
-        if($request->isMethod('POST')){
+        if ($request->isMethod('POST')) {
             $form->handleRequest($request);
 
             if ($form->isValid()) {
@@ -61,10 +79,29 @@ class RepairOrderController extends Controller
         $entity = $em->getRepository('AppBundle:RepairOrder')->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find RepairOrder entity.');
+            try {
+                throw $this->createNotFoundException('Unable to find RepairOrder entity.');
+            } catch (NotFoundHttpException $ex) {
+                return $this->render('default/index.html.twig', array(
+                        'errorMessages' => [
+                            $ex->getMessage()
+                        ]
+                    )
+                );
+            }
+
         }
 
-        $this->denyAccessUnlessGranted('view', $entity, 'Unauthorized access!');
+        try {
+            $this->denyAccessUnlessGranted('view', $entity, 'Access denied!');
+        } catch (AccessDeniedException $ex) {
+            return $this->render('default/index.html.twig', array(
+                    'errorMessages' => [
+                        $ex->getMessage()
+                    ]
+                )
+            );
+        }
         return $this->render('AppBundle:RepairOrder:show.html.twig', array(
                 'entity' => $entity
             )
@@ -97,7 +134,7 @@ class RepairOrderController extends Controller
             'method' => 'PUT',
         ));
         $editForm->add('submit', 'submit', array('label' => 'Update'));
-        if($request->isMethod('PUT')){
+        if ($request->isMethod('PUT')) {
             $editForm->handleRequest($request);
 
             if ($editForm->isValid()) {
@@ -120,34 +157,12 @@ class RepairOrderController extends Controller
         );
     }
 
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
-        $deleteForm = $this->createDeleteForm($id);
-
-        return $this->render('AppBundle:RepairOrder:delete.html.twig', array(
-                'delete_form' => $deleteForm->createView()
-            )
-        );
-    }
-
-    private function createDeleteForm($id)
-    {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('repairorder_remove', array(
-                'id' => $id
-            )))
-            ->setMethod('POST')
-            ->add('submit', 'submit', array('label' => 'Delete'))
-            ->getForm();
-    }
-
-    public function removeAction(Request $request, $id)
-    {
-        $deleteForm = $this->createDeleteForm($id);
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('AppBundle:RepairOrder')->find($id);
         try {
-            $this->denyAccessUnlessGranted('edit', $entity, 'Access denied to edit orders!');
+            $this->denyAccessUnlessGranted('delete', $entity, 'Access denied to delete.');
         } catch (AccessDeniedException $ex) {
             return $this->render('default/index.html.twig', array(
                     'errorMessages' => [
@@ -156,19 +171,44 @@ class RepairOrderController extends Controller
                 )
             );
         }
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find RepairOrder entity.');
+
+        $deleteForm = $this->createFormBuilder()
+            ->setAction($this->generateUrl('repairorder_delete', array(
+                'id' => $id
+            )))
+            ->setMethod('POST')
+            ->add('submit', 'submit', array('label' => 'Delete'))
+            ->getForm();
+        if ($request->isMethod('POST')) {
+            try {
+                $this->denyAccessUnlessGranted('edit', $entity, 'Access denied to edit orders!');
+            } catch (AccessDeniedException $ex) {
+                return $this->render('default/index.html.twig', array(
+                        'errorMessages' => [
+                            $ex->getMessage()
+                        ]
+                    )
+                );
+            }
+            if (!$entity) {
+                throw $this->createNotFoundException('Unable to find RepairOrder entity.');
+            }
+
+            $deleteForm->handleRequest($request);
+            if ($deleteForm->isValid()) {
+                $em->remove($entity);
+                $em->flush();
+
+                return $this->redirect($this->generateUrl('repair_orders'));
+            }
+            return $this->render('AppBundle:RepairOrder:delete.html.twig', array(
+                    'deleteForm' => $deleteForm->createView()
+                )
+            );
         }
 
-        $deleteForm->handleRequest($request);
-        if ($deleteForm->isValid()) {
-            $em->remove($entity);
-            $em->flush();
-
-            return $this->redirect($this->generateUrl('repair_orders'));
-        }
         return $this->render('AppBundle:RepairOrder:delete.html.twig', array(
-                'deleteForm' => $deleteForm->createView()
+                'delete_form' => $deleteForm->createView()
             )
         );
     }
